requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
cattrs==24.1.2
cdktf==0.20.11
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
constructs==10.4.2
distro==1.9.0
docker==7.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
jsii==1.106.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.60.0
orderly-set==5.2.3
publication==0.0.3
pydantic==2.10.5
pydantic_core==2.27.2
pydash==7.0.6
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2022.6
PyYAML==6.0.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
tomli==2.0.2
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.3.0
vellum-ai==0.13.7
-e git+https://github.com/alexfosterinvisible/little-things.git@9ca6997f86923749c6780e1fec4f3851da544a04#egg=vellum_quick_test&subdirectory=vellum-quick-test
